Update to BETA 10.13.5 (4856)

+++ sources/org/telegram/ui/Cells/MaxFileSizeCell.java
@@ -18,6 +18,7 @@ import org.telegram.messenger.R;
 import org.telegram.ui.ActionBar.Theme;
 import org.telegram.ui.Components.LayoutHelper;
 import org.telegram.ui.Components.SeekBarView;
+
 public class MaxFileSizeCell extends FrameLayout {
     private long currentSize;
     private SeekBarView seekBarView;
@@ -83,7 +84,7 @@ public class MaxFileSizeCell extends FrameLayout {
                 } else {
                     f -= 0.25f;
                     if (f < 0.25f) {
-                        f2 = (float) FileLoaderPriorityQueue.PRIORITY_VALUE_MAX;
+                        f2 = FileLoaderPriorityQueue.PRIORITY_VALUE_MAX;
                         f3 = 9437184.0f;
                     } else {
                         f -= 0.25f;
