+++ sources/org/telegram/ui/Components/Paint/RenderState.java
@@ -4,6 +4,7 @@ import android.graphics.PointF;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import org.telegram.messenger.LiteMode;
+
 public class RenderState {
     private int allocatedCount;
     public float alpha;
@@ -56,7 +57,7 @@ public class RenderState {
         if (this.buffer != null) {
             this.buffer = null;
         }
-        int max = Math.max(this.allocatedCount * 2, (int) LiteMode.FLAG_CHAT_BLUR);
+        int max = Math.max(this.allocatedCount * 2, LiteMode.FLAG_CHAT_BLUR);
         this.allocatedCount = max;
         ByteBuffer allocateDirect = ByteBuffer.allocateDirect(max * 5 * 4);
         this.buffer = allocateDirect;
